requirements.txt
Django==1.11.6
django-ckeditor==5.3.1
django-crispy-forms==1.7.0
